requirements.txt
python-docx==0.8.11
Flask==2.0.1
Werkzeug==2.0.1 